+++ src/RFC4180Field.php
@@ -50,10 +50,7 @@ class RFC4180Field extends php_user_filter
     public $filtername;
 
     /**
-     * Contents of the params parameter passed to stream_filter_append
-     * or stream_filter_prepend functions.
-     *
-     * @var mixed
+     * @var mixed value passed to passed to stream_filter_append or stream_filter_prepend functions.
      */
     public $params;