+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 All Notable changes to `Csv` will be documented in this file
+
 ## Next - TBD
 
 ### Added